healthGuideAPI/requirements.txt
asgiref==3.7.2
cmake==3.27.5
Django==4.2.5
djangorestframework==3.14.0
dlib @ file:///C:/Users/YENG%20SABS/Desktop/combo/dlib/dlib-19.22.99-cp310-cp310-win_amd64.whl#sha256=f2181b6724669fb0147c9ebc326d72b94de41bced1812eff5df4b133b5d0b575
Pillow==10.0.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
